+++ flutter_osm_interface/lib/src/channel/osm_method_channel.dart
@@ -27,7 +27,8 @@ class MethodChannelOSM extends MobileOSMPlatform {
 
   // Returns a filtered view of the events in the _controller, by mapId.
   Stream<EventOSM> _events(int mapId) =>
-      _streamController.stream.where((event) => event.mapId == mapId) as Stream<EventOSM>;
+      _streamController.stream.where((event) => event.mapId == mapId)
+          as Stream<EventOSM>;
 
   @override
   Future<void> init(int idOSMMap) async {
@@ -35,7 +36,8 @@ class MethodChannelOSM extends MobileOSMPlatform {
       if (_streamController.isClosed) {
         _streamController = StreamController<EventOSM>.broadcast();
       }
-      _channels[idOSMMap] = MethodChannel('plugins.dali.hamza/osmview_$idOSMMap');
+      _channels[idOSMMap] =
+          MethodChannel('plugins.dali.hamza/osmview_$idOSMMap');
       setGeoPointHandler(idOSMMap);
     }
     /*if (!_eventsChannels.containsKey(idOSMMap)) {
@@ -97,7 +99,8 @@ class MethodChannelOSM extends MobileOSMPlatform {
           break;
         case "receiveSinglePress":
           final result = call.arguments;
-          _streamController.add(SingleTapEvent(idMap, GeoPoint.fromMap(result)));
+          _streamController
+              .add(SingleTapEvent(idMap, GeoPoint.fromMap(result)));
           break;
         case "receiveGeoPoint":
           final result = call.arguments;
@@ -105,11 +108,13 @@ class MethodChannelOSM extends MobileOSMPlatform {
           break;
         case "receiveUserLocation":
           final result = call.arguments;
-          _streamController.add(UserLocationEvent(idMap, GeoPoint.fromMap(result)));
+          _streamController
+              .add(UserLocationEvent(idMap, GeoPoint.fromMap(result)));
           break;
         case "receiveRegionIsChanging":
           final result = call.arguments;
-          _streamController.add(RegionIsChangingEvent(idMap, Region.fromMap(result)));
+          _streamController
+              .add(RegionIsChangingEvent(idMap, Region.fromMap(result)));
           break;
       }
       return true;
@@ -150,7 +155,8 @@ class MethodChannelOSM extends MobileOSMPlatform {
   @override
   Future<GeoPoint> myLocation(int idMap) async {
     try {
-      Map<String, dynamic> map = (await (_channels[idMap]!.invokeMapMethod("user#position")))!;
+      Map<String, dynamic> map =
+          (await (_channels[idMap]!.invokeMapMethod("user#position")))!;
       return GeoPoint(latitude: map["lat"], longitude: map["lon"]);
     } on PlatformException catch (e) {
       throw GeoPointException(msg: e.message);
@@ -168,14 +174,10 @@ class MethodChannelOSM extends MobileOSMPlatform {
 
   @override
   Future<void> customMarker(int idOSM, GlobalKey? globalKey) async {
-    Uint8List icon = await _capturePng(globalKey!);
-    dynamic args = icon;
-    if (defaultTargetPlatform == TargetPlatform.iOS) {
-      var base64Str = base64.encode(icon);
-      args = base64Str;
-    }
+    final icon = await _capturePng(globalKey!);
+   
 
-    await _channels[idOSM]?.invokeMethod("marker#icon", args);
+    await _channels[idOSM]?.invokeMethod("marker#icon", icon);
   }
 
   @override
@@ -186,14 +188,11 @@ class MethodChannelOSM extends MobileOSMPlatform {
     bool refresh = false,
   }) async {
     if (globalKey?.currentContext != null) {
-      Uint8List icon = await _capturePng(globalKey!);
-      String iconIOS = "";
-      if (Platform.isIOS) {
-        iconIOS = icon.convertToString();
-      }
+      final icon = await _capturePng(globalKey!);
+
       var args = {
         "id": id,
-        "bitmap": Platform.isIOS ? iconIOS : icon,
+        "bitmap": icon,
         "refresh": refresh,
       };
 
@@ -233,7 +232,8 @@ class MethodChannelOSM extends MobileOSMPlatform {
 
     /// add middle point that will pass through it
     if (interestPoints != null && interestPoints.isNotEmpty) {
-      args.addAll({"middlePoints": interestPoints.map((e) => e.toMap()).toList()});
+      args.addAll(
+          {"middlePoints": interestPoints.map((e) => e.toMap()).toList()});
     }
 
     args.addAll(roadOption.toMap());
@@ -265,15 +265,15 @@ class MethodChannelOSM extends MobileOSMPlatform {
     Map args = {};
     if (key != null) {
       bitmap = await _capturePng(key);
-      args.addAll({"icon": Platform.isIOS ? bitmap.convertToString() : bitmap});
+      args.addAll({"icon": bitmap});
     }
     if (imageURL.isNotEmpty) {
       args.addAll({"imageURL": imageURL});
     }
 
     try {
-      Map<String, dynamic>? map =
-          (await (_channels[idOSM]?.invokeMapMethod("user#pickPosition", args)));
+      Map<String, dynamic>? map = (await (_channels[idOSM]
+          ?.invokeMapMethod("user#pickPosition", args)));
       return GeoPoint(latitude: map!["lat"], longitude: map["lon"]);
     } on PlatformException catch (e) {
       throw GeoPointException(msg: e.message);
@@ -287,7 +287,8 @@ class MethodChannelOSM extends MobileOSMPlatform {
 
   @override
   Future<void> removePosition(int idOSM, GeoPoint p) async {
-    await _channels[idOSM]?.invokeMethod("user#removeMarkerPosition", p.toMap());
+    await _channels[idOSM]
+        ?.invokeMethod("user#removeMarkerPosition", p.toMap());
   }
 
   @override
@@ -318,16 +319,16 @@ class MethodChannelOSM extends MobileOSMPlatform {
     final endKey = keys.last!;
     Map<String, dynamic> bitmaps = {};
     if (startKey.currentContext != null) {
-      Uint8List marker = await _capturePng(startKey);
-      bitmaps.putIfAbsent("START", () => Platform.isIOS ? marker.convertToString() : marker);
+      final marker = await _capturePng(startKey);
+      bitmaps.putIfAbsent("START", () => marker);
     }
     if (endKey.currentContext != null) {
-      Uint8List marker = await _capturePng(endKey);
-      bitmaps.putIfAbsent("END", () => Platform.isIOS ? marker.convertToString() : marker);
+      final marker = await _capturePng(endKey);
+      bitmaps.putIfAbsent("END", () => marker);
     }
     if (middleKey.currentContext != null) {
-      Uint8List marker = await _capturePng(middleKey);
-      bitmaps.putIfAbsent("MIDDLE", () => Platform.isIOS ? marker.convertToString() : marker);
+      final marker = await _capturePng(middleKey);
+      bitmaps.putIfAbsent("MIDDLE", () => marker);
     }
     await _channels[idOSM]?.invokeMethod("road#markers", bitmaps);
   }
@@ -352,15 +353,19 @@ class MethodChannelOSM extends MobileOSMPlatform {
     }
   }
 
-  Future<Uint8List> _capturePng(GlobalKey globalKey) async {
+  Future<dynamic> _capturePng(GlobalKey globalKey) async {
     if (globalKey.currentContext == null) {
       throw Exception("Error to draw you custom icon");
     }
     RenderRepaintBoundary boundary =
         globalKey.currentContext!.findRenderObject() as RenderRepaintBoundary;
     ui.Image image = await boundary.toImage();
-    ByteData byteData = (await (image.toByteData(format: ui.ImageByteFormat.png)))!;
+    ByteData byteData =
+        (await (image.toByteData(format: ui.ImageByteFormat.png)))!;
     Uint8List pngBytes = byteData.buffer.asUint8List();
+    if (defaultTargetPlatform == TargetPlatform.iOS) {
+      return pngBytes.convertToString();
+    }
     return pngBytes;
   }
 
@@ -410,7 +415,8 @@ class MethodChannelOSM extends MobileOSMPlatform {
 
   @override
   Future<GeoPoint> selectAdvancedPositionPicker(int idOSM) async {
-    Map mGeoPoint = (await (_channels[idOSM]?.invokeMapMethod("confirm#advanced#selection")))!;
+    Map mGeoPoint = (await (_channels[idOSM]
+        ?.invokeMapMethod("confirm#advanced#selection")))!;
     return GeoPoint.fromMap(mGeoPoint);
   }
 
@@ -450,8 +456,8 @@ class MethodChannelOSM extends MobileOSMPlatform {
   @override
   Future<GeoPoint> getPositionOnlyAdvancedPositionPicker(int idOSM) async {
     try {
-      Map? mGeoPoint =
-          (await (_channels[idOSM]?.invokeMapMethod("get#position#advanced#selection")));
+      Map? mGeoPoint = (await (_channels[idOSM]
+          ?.invokeMapMethod("get#position#advanced#selection")));
       return GeoPoint.fromMap(mGeoPoint!);
     } on Exception catch (e) {
       throw Exception(e);
@@ -497,8 +503,8 @@ class MethodChannelOSM extends MobileOSMPlatform {
       data["interestPoints"] = await interestPoints.encodedToString();
       if (keyIconForInterestPoints != null) {
         try {
-          final Uint8List bytes = await _capturePng(keyIconForInterestPoints);
-          data["iconInterestPoints"] = Platform.isIOS ? bytes.convertToString() : bytes;
+          data["iconInterestPoints"] =
+              await _capturePng(keyIconForInterestPoints);
         } catch (e) {}
       }
     }
@@ -525,14 +531,9 @@ class MethodChannelOSM extends MobileOSMPlatform {
     int idMap,
     GlobalKey key,
   ) async {
-    Uint8List icon = await _capturePng(key);
+    final icon = await _capturePng(key);
 
-    if (defaultTargetPlatform == TargetPlatform.iOS) {
-      var base64Str = base64.encode(icon);
-      await _channels[idMap]!.invokeMethod("advancedPicker#marker#icon", base64Str);
-    } else {
-      await _channels[idMap]!.invokeMethod("advancedPicker#marker#icon", icon);
-    }
+    await _channels[idMap]!.invokeMethod("advancedPicker#marker#icon", icon);
   }
 
   @override
@@ -556,17 +557,11 @@ class MethodChannelOSM extends MobileOSMPlatform {
     GlobalKey personGlobalKey,
     GlobalKey directionArrowGlobalKey,
   ) async {
-    Uint8List iconPerson = await _capturePng(personGlobalKey);
-    Uint8List iconArrowDirection = await _capturePng(directionArrowGlobalKey);
+    final iconPerson = await _capturePng(personGlobalKey);
+    final iconArrowDirection = await _capturePng(directionArrowGlobalKey);
     HashMap<String, dynamic> args = HashMap();
     args["personIcon"] = iconPerson;
     args["arrowDirectionIcon"] = iconArrowDirection;
-    if (defaultTargetPlatform == TargetPlatform.iOS) {
-      var base64StrPerson = base64.encode(iconPerson);
-      var base64StrArrowDirection = base64.encode(iconPerson);
-      args["personIcon"] = base64StrPerson;
-      args["arrowDirectionIcon"] = base64StrArrowDirection;
-    }
 
     await _channels[idOSM]?.invokeMethod("user#locationMarkers", args);
   }
@@ -580,7 +575,7 @@ class MethodChannelOSM extends MobileOSMPlatform {
     Map<String, dynamic> args = {"point": p.toMap()};
     if (globalKeyIcon != null) {
       var icon = await _capturePng(globalKeyIcon);
-      args["icon"] = Platform.isIOS ? icon.convertToString() : icon;
+      args["icon"] = icon;
     }
 
     await _channels[idOSM]?.invokeMethod("add#Marker", args);
@@ -649,8 +644,7 @@ class MethodChannelOSM extends MobileOSMPlatform {
     GlobalKey<State<StatefulWidget>> globalKeyIcon,
   ) async {
     Map<String, dynamic> args = {"point": point.toMap()};
-    var icon = await _capturePng(globalKeyIcon);
-    args["icon"] = Platform.isIOS ? icon.convertToString() : icon;
+    args["icon"] = await _capturePng(globalKeyIcon);
     try {
       await _channels[idOSM]?.invokeMethod("update#Marker", args);
     } on PlatformException catch (e) {
@@ -672,7 +666,8 @@ class MethodChannelOSM extends MobileOSMPlatform {
     MultiRoadOption commonRoadOption = const MultiRoadOption.empty(),
   }) async {
     final args = configs.toListMap(commonRoadOption: commonRoadOption);
-    final List result = (await _channels[idOSM]?.invokeListMethod("draw#multi#road", args))!;
+    final List result =
+        (await _channels[idOSM]?.invokeListMethod("draw#multi#road", args))!;
     final List<Map<String, dynamic>> mapRoadInfo =
         result.map((e) => Map<String, dynamic>.from(e)).toList();
     return mapRoadInfo.map((e) => RoadInfo.fromMap(e)).toList();
@@ -696,8 +691,7 @@ class MethodChannelOSM extends MobileOSMPlatform {
       "old_location": oldLocation.toMap(),
     };
     if (globalKeyIcon != null) {
-      var icon = await _capturePng(globalKeyIcon);
-      args["new_icon"] = Platform.isIOS ? icon.convertToString() : icon;
+      args["new_icon"] = await _capturePng(globalKeyIcon);
     }
     await _channels[idOSM]!.invokeMethod("change#Marker", args);
   }
@@ -723,7 +717,7 @@ extension config on MethodChannelOSM {
 
   Future<void> initIosMap(int idOSM, GlobalKey key) async {
     await _channels[idOSM]?.invokeMethod("init#ios#map");
-    final icon = (await _capturePng(key)).convertToString();
+    final icon = await _capturePng(key);
     await _channels[idOSM]?.invokeMethod("setDefaultIOSIcon", icon);
   }
 }